refactor: rename Video.ts to VideoStore.ts< actions.ts to ActionTypes.ts
