+++ src/api/search-service.ts
@@ -1,4 +1,4 @@
-import { Video } from "../types/Video";
+import { Video } from "../types/VideoStore";
 import {
   YouTubeSearchResult,
   YouTubeVideo,
